requirements.txt
asgiref==3.6.0
boto3==1.34.35
botocore==1.34.35
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
cryptography==39.0.1
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.57.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-storages==1.14.2
environs==10.3.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
Markdown==3.5.2
markdown-it-py==3.0.0
marshmallow==3.20.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.1
pillow==10.2.0
psycopg==3.1.12
psycopg-binary==3.1.12
pycparser==2.21
pydot==2.0.0
Pygments==2.17.2
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.8.0
tzdata==2023.4
urllib3==1.26.14
whitenoise==6.5.0
